Refactorizat clasele celulelor si modificat generarea numelor
