+++ celluloid/src/main/java/com/celluloid/controller/AddSexualCell.java
@@ -1,0 +1,10 @@
+package com.celluloid.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(value = "new-sexual-cell")
+public class AddSexualCell {
+
+}
